+++ CHANGELOG.rst
@@ -1,8 +1,16 @@
 CHANGELOG
 =========
 
+1.10.1
+------
+
+- fix: address a cross-site scripting vulnerability where member and book range
+  search inputs were displaying raw query request values even when invalid/unsafe
+- Update webpack config to use `contenthash` for outputs since `hash` is deprecated
+
+
 1.10
----
+----
 
 - Site Menus and taglines and Homepage tagline now editable in Wagtail
 - Enhancements to Wagtail editorial essays :